agent_framework/requirements.txt
torch>=2.0.0
transformers>=4.37.0
peft>=0.7.0
datasets>=2.14.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99
protobuf>=3.20.0
pyyaml>=6.0
sentence-transformers>=2.2.0
numpy>=1.24.0

